move pattern to a separate module
